Bug fixes and UI updates from Kirsten
Fixed: small bug when scoring with keyboard shortcuts in first sentence in continuous noise.
Fixed: setting correct target/masker levels when using a randomization type. If result type = SRT, it only reads the targetLevel from the first sentence in block as start, but afterwards calculates automatically the SNR from the maskerLevel and the nextSignalToNoise. For result type = Percentage correct, however, reads values directly from csv. (Obs. it's the user's obligation to make sure they upload a valid file).
Added: Target and Masker dropdown menus are updated automatically in the Startup fcn depending on the available files inside the corresponding folders (same behavior as the list type).
Added: 'drawnow limitrate' for the while loop in playContinuousMaskerAndMixOnTheFly.
Added: when using advanced mode, if uploading a randomization file defined by lists, the list number is also saved in the result file - and not only the sentence number (same behavior as in QuickStart).